refactor(plugin): move shared host-factory types to PluginServiceTypes.ts
- Relocate LoadedPlugin, ValidateFn, WorkspaceWorktreeEvent, FsRootClass,
  and ExpandedFsPath into a new electron/services/plugin/PluginServiceTypes.ts
  instead of exporting them from PluginService.ts purely for
  PluginHostFactory.ts's benefit — keeps PluginService.ts's export surface
  unchanged
- Fix an inaccurate comment (5 disposer call sites, not ~6) and a stale
  'constructed once' note in PluginHostFactory.ts
